feat: new plugin hashicorp-vault: rename plugin name to hashicorp-vault
