+++ cmd/plugin/hashicorp-vault/main.go
@@ -1,34 +1,34 @@
 package main
 
 import (
-	"github.com/devstream-io/devstream/internal/pkg/plugin/vault"
+	"github.com/devstream-io/devstream/internal/pkg/plugin/hashicorpvault"
 	"github.com/devstream-io/devstream/pkg/util/log"
 )
 
 // NAME is the name of this DevStream plugin.
-const NAME = "vault"
+const NAME = "hashicorp-vault"
 
 // Plugin is the type used by DevStream core. It's a string.
 type Plugin string
 
 // Create implements the create of vault.
 func (p Plugin) Create(options map[string]interface{}) (map[string]interface{}, error) {
-	return vault.Create(options)
+	return hashicorpvault.Create(options)
 }
 
 // Update implements the update of vault.
 func (p Plugin) Update(options map[string]interface{}) (map[string]interface{}, error) {
-	return vault.Update(options)
+	return hashicorpvault.Update(options)
 }
 
 // Delete implements the delete of vault.
 func (p Plugin) Delete(options map[string]interface{}) (bool, error) {
-	return vault.Delete(options)
+	return hashicorpvault.Delete(options)
 }
 
 // Read implements the read of vault.
 func (p Plugin) Read(options map[string]interface{}) (map[string]interface{}, error) {
-	return vault.Read(options)
+	return hashicorpvault.Read(options)
 }
 
 // DevStreamPlugin is the exported variable used by the DevStream core.
