+++ internal/pkg/plugin/hashicorpvault/delete.go
@@ -1,4 +1,4 @@
-package vault
+package hashicorpvault
 
 import (
 	"fmt"
