+++ internal/pkg/plugin/hashicorpvault/update.go
@@ -1,4 +1,4 @@
-package vault
+package hashicorpvault
 
 import (
 	"fmt"
