+++ internal/pkg/show/config/config.go
@@ -24,7 +24,7 @@ var pluginDefaultConfigs = map[string]string{
 	"openldap":                       plugin.OpenldapDefaultConfig,
 	"trello-github-integ":            plugin.TrelloGithubDefaultConfig,
 	"trello":                         plugin.TrelloDefaultConfig,
-	"vault":                          plugin.VaultDefaultConfig,
+	"hashicorp-vault":                plugin.VaultDefaultConfig,
 }
 
 func Show() error {
