+++ internal/pkg/show/config/plugin/hashicorp-vault.go
